ZigBee: Allow a different set of attributes for ZCL client and server
Some clusters, for example the Smart Energy Metering cluster,
requires a different set of attributes for the ZCL client and
server but with overlapping IDs.

This change allows to specify a different set of attributes for
the ZCL client.

To avoid breaking existing clusters that might have client attributes
the same set of attributes are used for server and client when
zbee_zcl_init_cluster is called.

This new client attribute set is used in the Smart Energy Metering
cluster in this commit.

Change-Id: Ie25ad746dac1ccbb233ae8975ef9047d3fc6a170
Reviewed-on: https://code.wireshark.org/review/26694
Reviewed-by: Martin Boye Petersen <martinboyepetersen@gmail.com>
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Paul Zander <p.j.zander@lighting.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>