If we're reading from a string, don't fclose yyin.
yyin is initialized to stdin.  When we're reading from files, we set it
so that it points to the FILE from which we're reading, but when we're
reading from a string, we don't set it, leaving it to point to stdin.

This means that, just as the "read from the input" routine has to be set
differently when reading from a file or a string, the "close the current
input" routine has to be set differently as well.

Bug: 14577
Change-Id: Ie05880775612867e9037ace2de0cd0a0dd2fabb5
Reviewed-on: https://code.wireshark.org/review/26719
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 9d87f607ee2ecefaff71f2ab3f2dc3d2dc185399)
Reviewed-on: https://code.wireshark.org/review/26721