Get rid of get_opaque_length() and unneeded length arguments.
In most calls to get_opaque_length(), the length is a constant, so just
directly fetch the length value.  In the calls in dissect_nodeid_list()
and dissect_kindid_list(), the length is an argument to the routine
making the call, and *those* arguments are, for each of the routines,
the same constant for all calls to the routine in question, so just
directly fetch the length value in those routines.

This lets us avoid checks for a valid length-of-length, so we don't need
to have the length-of-length be a signed value with -1 meaning
"invalid".  That's good, because nothing was actually *checking* for an
invalid length-of-length.

Change-Id: I58264c133977266f3214d6e4ca361f71ecc0b69a
Reviewed-on: https://code.wireshark.org/review/26842
Reviewed-by: Guy Harris <guy@alum.mit.edu>