Builds require CMake, so always install it.
Don't have an option not to install it.  (If we want to avoid installing
it when it's already been installed by something other than this script,
we should check for its existence before installing, and skip the
installation step for it.)

Get rid of the instructions for autotools builds; always show the
instructions for CMake builds.

We wouldn't get to the uninstall stage if it hadn't been installed by
this script; remove the comment asking about that.

Change-Id: I276ee96bf955ef4ff33dea87bc27c21111301ea0
Reviewed-on: https://code.wireshark.org/review/27540
Reviewed-by: Guy Harris <guy@alum.mit.edu>